requirements.txt
Django>=1.5.1,<1.6
SQLAlchemy==0.8.0
amqp==1.0.10
anyjson==0.3.3
billiard==2.7.3.23
celery==3.0.17
coverage==3.7.1
cubes==0.10.2
django-celery==3.0.17
django-debug-toolbar==0.11.0
django-extensions==1.3.2
django-tastypie==0.11.0
jsonfield==1.0.0
kombu==2.5.8
lxml==3.3.6
mimeparse==0.1.3
mock==1.0.1
psycopg2>=2.4.6
python-dateutil==1.5
pytz==2014.9
requests==2.4.3
selenium==2.32.0
simplejson==3.1.3
six==1.3.0
statistics==1.0.3.5
vecnet.emod==0.6.0
vecnet.simulation==0.2
vecnet.openmalaria==0.9.1.dev5
wsgiref==0.1.2
-e djorm-ext-pgbytea
